api: move vim25.Client.UseServiceVersion to soap.Client
This method only changes fields in soap.Client
Moving it allows UseServiceVersion to be called after NewServiceVersion.
